refactor: move common attribute of file-based data node to _FileDataNodeMixin
