+++ taipy/core/data/json.py
@@ -60,9 +60,6 @@ class JSONDataNode(DataNode, _FileDataNodeMixin):
     """
 
     __STORAGE_TYPE = "json"
-    _DEFAULT_DATA_KEY = "default_data"
-    _DEFAULT_PATH_KEY = "default_path"
-    _PATH_KEY = "path"
     __ENCODING_KEY = "encoding"
     _ENCODER_KEY = "encoder"
     _DECODER_KEY = "decoder"
@@ -94,8 +91,6 @@ class JSONDataNode(DataNode, _FileDataNodeMixin):
 
         default_value = properties.pop(self._DEFAULT_DATA_KEY, None)
         _FileDataNodeMixin.__init__(self, properties)
-        properties[self._IS_GENERATED_KEY] = self._is_generated
-        properties[self._PATH_KEY] = self._path
 
         DataNode.__init__(
             self,
@@ -134,17 +129,6 @@ class JSONDataNode(DataNode, _FileDataNodeMixin):
     @classmethod
     def storage_type(cls) -> str:
         return cls.__STORAGE_TYPE
-
-    @property  # type: ignore
-    @_self_reload(DataNode._MANAGER_NAME)
-    def path(self):
-        return self._path
-
-    @path.setter
-    def path(self, value):
-        self._path = value
-        self.properties[self._PATH_KEY] = value
-        self.properties[self._IS_GENERATED_KEY] = False
 
     @property  # type: ignore
     @_self_reload(DataNode._MANAGER_NAME)
